take book logic into book_options
