+++ app.rb
@@ -1,99 +1,48 @@
-require './book'
+require './book_option'
 require './rental'
-require './student'
-require './teacher'
+require './person_option'
 
 class App
   def initialize(main_call)
     @main_call = main_call
-    @people = []
-    @books = []
+    @person_option = PersonOption.new
+    @book_options = BookOption.new
     @rentals_list = []
   end
 
   def list_books
-    if @books.empty?
-      puts 'No books available'
-    else
-      puts 'List of books:'
-      @books.each { |book| puts "Title: #{book.title}, Author: #{book.author}" }
-    end
+    @book_options.list_books
     @main_call.display_menu
   end
 
   def list_people
-    if @people.empty?
-      puts 'No people available'
-    else
-      puts 'List of people:'
-      @people.each do |person|
-        puts "[#{person.class} Name: #{person.name}, ID: #{person.id}", "Age: #{person.age}]"
-      end
-    end
+    @person_option.list_people
     @main_call.display_menu
   end
 
   def create_person
-    puts 'Do you want to create a student (1) or a teacher (2)? [Input the number]:'
-    person_type = gets.chomp.to_i
-    case person_type
-    when 1
-      create_student
-    when 2
-      create_teacher
-    else
-      puts 'Invalid option'
-    end
+    @person_option.create_person
     @main_call.display_menu
   end
 
-  def create_student
-    puts 'Age:'
-    age = gets.chomp.to_i
-    puts 'Name:'
-    name = gets.chomp
-    puts 'Has parent permission? [Y/N]:'
-    parent_permission = gets.chomp.upcase == 'Y'
-    @people.push(Student.new(age, name, parent_permission: parent_permission))
-    puts
-    puts "Person #{name} was created"
-  end
-
-  def create_teacher
-    puts 'Age:'
-    age = gets.chomp.to_i
-    puts 'Name:'
-    name = gets.chomp
-    puts 'Specialization:'
-    specialization = gets.chomp
-    @people.push(Teacher.new(age, specialization, name))
-    puts
-    puts "Person #{name} was created"
-  end
-
   def create_book
-    puts 'Title:'
-    title = gets.chomp
-    puts 'Author:'
-    author = gets.chomp
-    @books.push(Book.new(title, author))
-    puts "Book #{title} was created"
+    @book_options.create_book
     @main_call.display_menu
   end
 
   def create_rental
     puts 'Select a book from the following list by number'
-    @books.each_with_index { |book, index| puts "#{index} Title: #{book.title}, Author: #{book.author}" }
+    @book_options.books.each_with_index { |book, index| puts "#{index} Title: #{book.title}, Author: #{book.author}" }
     book_index = gets.chomp.to_i
     puts
     puts 'Select a person from the following list by number (not ID)'
-    @people.each_with_index do |person, index|
+    @person_option.people.each_with_index do |person, index|
       puts "#{index} [#{person.class}] Name: #{person.name}, ID: #{person.id}", "Age: #{person.age}"
     end
     person_index = gets.chomp.to_i
     puts 'Date:'
     date = gets.chomp
-    @rentals_list.push(Rental.new(date, @books[book_index], @people[person_index]))
+    @rentals_list.push(Rental.new(date, @book_options.books[book_index], @person_option.people[person_index]))
     puts 'Rental created successfully'
     @main_call.display_menu
   end
